Raw data management
Opens .h5 files exported from the MCS data manager. Can manage the data by downsampling, and renaming columns. Data is organized as data frames and saved as .csv files for further analysis with the SWP analyzer in R, or in general to analyzed in R or python.. or to that extent any othe language.